homelessAPI/requirements.txt
appdirs==1.4.0
coreapi==2.1.1
coverage==4.3.4
Django==1.10.5
django-nose==1.4.4
django-rest-swagger==2.1.1
djangorestframework==3.5.3
itypes==1.1.0
nose==1.3.7
openapi-codec==1.2.1
packaging==16.8
psycopg2==2.6.2
pyparsing==2.1.10
requests==2.13.0
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
